requirements.txt
lxml
opencc_python_reimplemented
requests
PyQt5
PyQt5-tools
PyYAML
nicegui
ttkbootstrap